requirements.txt
attrs==23.1.0
boto3>=1.34.17
botocore>=1.34.17
cryptography>=42.0.5
dnspython>=2.1.0
google-api-core>=2.4.0
google-api-python-client>=2.34.0
google-auth>=2.3.3
google-auth-httplib2>=0.1.0
googleapis-common-protos>=1.54.0
google-cloud>=0.34.0
google-cloud-compute>=1.6.1
google-cloud-storage>=2.10.0
google-cloud-dns>=0.35.0
google-cloud-network-management>=1.5.4
Jinja2>=3.1.3
passlib>=1.7.4
pycryptodome>=3.20.0
pytz>=2021.3
pyvmomi>=8.0.0.1.1
requests>=2.31.0
urllib3>=1.26.16
azure-common>=1.1.28
azure-core>=1.29.6
azure-mgmt-resource>=22.0.0
azure-identity>=1.12.0
azure.mgmt.network>=25.1.0
azure.mgmt.compute>=29.0.0
azure-mgmt-core>=1.4.0
azure.mgmt.storage>=21.0.0
azure-mgmt-dns>=8.1.0
azure-mgmt-privatedns>=1.1.0
azure.mgmt.subscription>=3.1.1
ply>=3.11
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-rerunfailures==12.0
sqlite-utils~=3.11
docker>=5.0.3
paramiko==3.4.0
overrides>=7.4.0
bumpversion>=0.6.0
PyYAML>=5.1
rsa>=4.9
pywinrm>=0.4.3
aiohttp>=3.9.3
python-certifi-win32>=1.6.1
certifi>=2023.5.7
